feat: rename preview command to view, add CP dashboard link support
