+++ cmd/kubectl-testkube/commands/root.go
@@ -59,7 +59,7 @@ func init() {
 	RootCmd.AddCommand(NewPurgeCmd())
 	RootCmd.AddCommand(NewWatchCmd())
 	RootCmd.AddCommand(NewDashboardCmd())
-	RootCmd.AddCommand(NewPreviewCmd())
+	RootCmd.AddCommand(NewViewCmd())
 	RootCmd.AddCommand(NewMigrateCmd())
 	RootCmd.AddCommand(NewVersionCmd())
